requirements.txt
git+https://github.com/mobarski/ai-bricks.git
streamlit~=1.25.0
pypdf~=3.15.1
scikit-learn~=1.3.0
numpy
pycryptodome
boto3~=1.28.30
redis~=5.0.0
retry~=0.9.2

vertexai~=0.0.1
Pillow~=9.5.0
botocore~=1.31.30